refactor(dart): redesign nested field specs and schema hash

+++ dart/test/workspace_smoke_test.dart
@@ -23,8 +23,8 @@ import 'package:test/test.dart';
 void main() {
   test('workspace root can use the rewritten fory package', () {
     final fory = Fory();
-    final bytes = fory.serialize(Int32(42));
-    final roundTrip = fory.deserialize<Int32>(bytes);
-    expect(roundTrip, equals(Int32(42)));
+    final bytes = fory.serialize(42);
+    final roundTrip = fory.deserialize<int>(bytes);
+    expect(roundTrip, equals(42));
   });
 }
